Add a confirmation_tones option to bring back some of the old beeps
I kind of predicted this would happen - switch to a new workflow and the
first thing everyone asks for is to bring back the old one :)

This adds an option to do so - though, it is intentionally limited to
the shader reload and copy on mark confirmations, and does not bring
back any of the ini parse warning tones - nothing that needs the log
file checked to understand the meaning.

The option is disabled by default to encourage the new workflow be given
a fair go. Personally while I initially missed the confirmation tones,
it didn't take long before I was fine with their absence, and enjoying
the fact that I don't have to have my speakers up to an audible level or
headphones in to work on a fix, and that there is less stuttering when
reloading (since the beeps lock the game up for a moment).