Set ACL on newly created files to try to ensure we have access
This is related to the work to set ACLs on newly created directories to
ensure we can read back the files we create, but it addresses different
edge cases to ensure we can read any shaders we dump (still have no idea
how or why these arise).

This creates a utility routine we can call in place of _wfopen_s to set
security attributes on the file when it is created. In order to do this
we actually call CreateFile, then convert the file handle into a file
descriptor, and finally a file pointer / stream compatible with that
returned by _wfopen_s for use with formatted file print routines.

This utility routine is only for creating files at the moment. _wfopen_s
should continue to be used when reading a file, or appending to an
existing file. We can certainly make it more flexible, but I only want
these heroics to go as far as they need to to address real life windows
filesystem permission issues.

For example, say a game directory got these permissions:

    dss@Gideon ~/MGS_TPP
    $ icacls.exe .
    . BUILTIN\Users:(I)(F)
      BUILTIN\Users:(I)(OI)(CI)(IO)(F)
      NT AUTHORITY\SYSTEM:(I)(F)
      NT AUTHORITY\SYSTEM:(I)(OI)(CI)(IO)(F)
      BUILTIN\Administrators:(I)(F)
      BUILTIN\Administrators:(I)(OI)(CI)(IO)(F)
      NT AUTHORITY\Authenticated Users:(I)(M)
      NT AUTHORITY\Authenticated Users:(I)(OI)(CI)(IO)(M)
      BUILTIN\Users:(I)(OI)(CI)(IO)(GR,GE)

And ShaderFixes somehow managed to get these permissions:

   $ icacls.exe ShaderFixes/
   ShaderFixes/ NULL SID:(DENY)(Rc,S,REA,X,DC)
                GIDEON\dss:(F)
                NT AUTHORITY\Authenticated Users:(DENY)(W,DC)
                BUILTIN\Users:(DENY)(W,DC)
                GIDEON\None:(Rc,S,RA)
                NT AUTHORITY\Authenticated Users:(RX,W,DC)
                NT AUTHORITY\SYSTEM:(RX,W,DC)
                BUILTIN\Administrators:(RX,W,DC)
                BUILTIN\Users:(RX,W,DC)
                GIDEON\None:(DENY)(S,RD,REA,X)
                Everyone:(RX)
                NULL SID:(OI)(CI)(IO)(DENY)(Rc,S,REA,WEA,X,DC)
                CREATOR OWNER:(OI)(CI)(IO)(DENY)(S,RD,WD,AD,REA,WEA,X,DC)
                CREATOR OWNER:(OI)(CI)(IO)(D,Rc,WDAC,WO,RA,WA)
                CREATOR GROUP:(OI)(CI)(IO)(Rc,S,RA)
                NT AUTHORITY\Authenticated Users:(OI)(CI)(IO)(RX,W,DC)
                NT AUTHORITY\SYSTEM:(OI)(CI)(IO)(RX,W,DC)
                BUILTIN\Administrators:(OI)(CI)(IO)(RX,W,DC)
                BUILTIN\Users:(OI)(CI)(IO)(RX,W,DC)
                Everyone:(OI)(CI)(IO)(Rc,S,RA)

Then newly dumped shaders end up with these permissions:

    $ icacls.exe  000000001247b14e-vs_replace.txt
    000000001247b14e-vs_replace.txt NULL SID:(I)(DENY)(Rc,S,REA,WEA,X,DC)
                                    GIDEON\dss:(I)(DENY)(S,RD,WD,AD,REA,WEA,X,DC)
                                    GIDEON\dss:(I)(D,Rc,WDAC,WO,RA,WA)
                                    GIDEON\None:(I)(Rc,S,RA)
                                    NT AUTHORITY\Authenticated Users:(I)(RX,W,DC)
                                    NT AUTHORITY\SYSTEM:(I)(RX,W,DC)
                                    BUILTIN\Administrators:(I)(RX,W,DC)
                                    BUILTIN\Users:(I)(RX,W,DC)
                                    Everyone:(I)(Rc,S,RA)

Which doesn't allow 3DMigoto to read it, and 3DMigoto issues a warning as a result:

    dss@Gideon ~/MGS_TPP/ShaderFixes
    $ cat 000000001247b14e-vs_replace.txt
    cat: 000000001247b14e-vs_replace.txt: Permission denied

Following this change we now get these permissions instead:

    $ icacls.exe  000000001247b14e-vs_replace.txt
    000000001247b14e-vs_replace.txt Everyone:(OI)(CI)(RX)
                                    NT AUTHORITY\Authenticated Users:(OI)(CI)(F)
                                    BUILTIN\Administrators:(OI)(CI)(F)

Which allows us to read the file.