Implement upscaling on CreateSwapChainForHwnd path
This hooks up a similar upscaling path for games that use
CreateSwapChainForHwnd as for our regular CreateSwapChain path.

This is only lightly tested in a Unity game, and it's definitely doing
something - but I'm not certain if it is doing the right thing - Unity's
"resolution" setting does not seem to actually mean "resolution", so I
don't even know if the test is valid.