Implement resolution overrides in CreateSwapChainForHwnd path
This seems to more or less work as it does in the regular init path,
however it mostly just results in a smaller sized window, and with
full_screen=1 results in a corrupt render... though this is pretty
consistent with the overrides in the regular path as well, since it has
not resized any of the other full screen buffers, such as the depth
buffer.

This probably needs upscaling implemented in this path to be useful.