Refresh rate overrides: Remove full-screen check
Historically we have only forced the refresh rate when full-screen.  I
don't know if we ever had a good reason for that, but it complicates
forcing the refresh rate in games that start windowed and later switch
to full screen, depending on the order in which the game calls
ResizeTarget and SetFullscreenState.

UE4 does SetFullscreenState -> ResizeBuffers -> ResizeTarget, which
works with our current code since it is fullscreen by the time we force
the refresh rate in ResizeTarget, but Unity does ResizeTarget ->
SetFullscreenState -> ResizeBuffers, so it is still windowed when it
calls ResizeTarget and as such we miss the opportunity to force the
refresh.

According to MSDN, both are valid, though the UE4 way will generate one
extra WM_SIZE. That suggests that it is valid to set the refresh rate
while still in windowed mode and this should be safe - though I am
hesitant given someone had added a full screen check and maybe they had
a good reason I don't know about. If that does turn out to be the case,
we would need to go to further heroics to make every case work, and I'd
rather not if we don't need to - so, remove the full-screen check and
we will see how it fares in practice.

Some testing reveals that this allows the 3D TV Play green flicker bug
to resurface in Tekken 7, but only when the desktop resolution is set to
1280x720 60Hz 3D ahead of time and that is the same resolution the game
is using. Without this patch that scenario is broken anyway, as the game
still tries to use 120Hz and cannot engage 3D. Having the desktop
resolution set to 1920x1080 and the game resolution set to 1280x720 and
forcing the refresh rate to 60Hz with 3DMigoto works either way, so at
least this doesn't seem to make anything worse.