Bring back audible tones on shader/config reload failure
After trying out the new overlay notification scheme for some time it
has become apparent that sometimes when pressing reload my gaze may be
focussed on something I expect to change, and not necessarily looking at
the top of the screen to see a failure if I have made a typo or
something. This isn't too severe, but adds a moment or two of confusion
as to why nothing has changed, or a shader unexpectedly reverted until
the message is noticed.

On the suggestion of Bo3b, this brings back audible notifications for
failure only. A tone will be sounded if either shaders fail to reload or
the config parser issues a warning (max one tone each, so max two tones
if both shaders and the config is broken) which will serve to alert the
shaderhacker that something is wrong, and the overlay notification will
inform them of what is wrong.