Custom Shaders: Add "flags" option to override compile flags
We sometimes might need to override the flags we pass into the compiler
for various reasons, e.g. my buffer debug shader needs to be compiled
with IEEE strictness enabled to prevent the compiler from optimising out
the NaN handling. It may also sometimes be desirable to lower the
optimisation level on debug shaders to make the F10 reload faster and
potentially there might be other things we want to do from time to time,
and this gives us a way to do that.

Specifying any custom flags will revert the optimisation to level 1, but
the optimization_level3 flag can easily be specified as well to counter
that.