Custom Shaders: Add pre-processor defines for shader types
Custom shaders at the moment have to ship separate HLSL files for each
shader type, which increases the amount of files we need to ship.

This adds a preprocessor definition for each shader type, allowing us to
place multiple shaders of different types in a single file so that they
can all be shipped as one. This is superior as it also allows common
structure definitions to be used as the input/output of each shader
stage, and this technique is often used by real game engines in the
wild.