Software mouse shader: Switch to using a struct to pass values between shaders
This is to try to give an example of the preferred method to pass values
between shader stages, which is less error prone when they are in the
same file using the same structure definition. It's not a great example
since there is only SV_Position and one texcoord, but it will do.