Add HackerCreateSwapChain to avoid other hooks.
There is no known scenario where we need this, but it makes
sense to make this behavior match the same technique used
in CreateDevice.  This adds a shim layer to the external call of
CreateSwapChain, which calls HackerCreateSwapChain.  In
CreateDeviceAndSwapChain, we now directly call HackerCreateSwapChain, to avoid relying upon our hook, which itself may be hooked by others. It's no longer an
implicit call.

Improve logging for CreateDeviceAndSwapChain.  This call
and CreateSwapChain are key calls that we want to see
happen, so this makes them stand out more in the log.

Tested in Headlander, BatmanTT, Dishonored2, and works.