Call Hooked_CreateDXGIFactory during InitD311 to set hooks.
If a game calls CreateDeviceAndSwapChain, but has never created
a DXGIFactory, then our CreateSwapChain hook will be null, leading
to crasharoo.

This fleshes out the init sequence to include a call to Hooked_CreateDXGIFactory,
so that we can be certain that it has been called at least once, and
therefore has installed the hook on CreateSwapChain.

If it's already been called, no harm, because it will just create
a DXGIFactory that we then immediately Release.

It is specifically done *after* hD3D11 is initialized, so that we
don't loop on inits, because Hooked_CreateDXGIFactory will
call InitD311 in case *it* happens to be called first.  Having two
different and incompatible ways to create SwapChains totally sucks.