Properly handle Overlay delete.
We are using the context, device, and swapchain passed in...
but we never say that by calling AddRef.

This now calls AddRef on the components we are using to
make sure they are not prematurely deleted.

And adds the correct Release ops in the destructor.  And
actually calls Delete when the SwapChain is Released.

Not sure this matters much, have not seen negative impacts,
outside of FFXIV, but it's closer to correct.  Overlay throws an
exception during creation there, and this does not solve the
problem, but it was clear that the context and swapchain were
stale.