Refactor LoadBinaryShaders
This splits LoadBinaryShaders into three functions - one to set the
policy on which shaders are loaded, a second to actually load a shader,
and a helper function to retrieve the timestamp of a file by path.

The code is cleaned up a little bit, but this shouldn't have any
functional changes yet, but we will need this split to fix some bugs in
the way we handle cached shaders.