Fix using stale .bin files over newer .txt files on initial load
If a .bin file for a shader is found (either from HLSL or assembly), it
will be given priority over any corresponding .txt file on initial load,
regardless of which is newer.

In practice this means if a fix is shipped with cache_shaders=1 then
later is updated, any shaders that have been altered by the update will
not be used. While this can be avoided by using uninstall.bat before
installing the new fix, it is preferable to not have to rely on that.

A second problem is that if a shaderhacker is running with
cache_shaders=1 and edits and reloads a shader it will seem to work just
fine, but if they exit and restart the game 3DMigoto will load the old
stale cached shader without their edits instead.

This compares the timestamps on the .bin and .txt files, and only uses
the .bin file if it is newer or matches the .txt file.

If there is no .txt file, the .bin file will still be used. This is
largely for historical reasons, as we have allowed (but discouraged)
binary only fixes in the past, but more to the point this is an unusual
scenario for an end user - if they updated a fix with uninstall.bat it
should have removed any cached .bin files, and if they did not they
should have only added new files or updated existing ones, and should
not have been able to remove a .txt file but leave the .bin file... This
might happen to a shaderhacker, but that is user error and there's not
much we can do about it other than issue a warning.