ShaderRegex: Fix "no more memory" error
When we perform a regex substitution we don't know how large the
replaced shader will be in advance, but calculating that is costly, so
we make a generous guess for the first attempt and if we underestimated
we are supposed to retry once we know the correct size. The guess was
generous enough that it has taken us until now to discover that the
retry code path was broken due to testing for an incorrect return code,
causing us to return "WARNING: regex replace error: no more memory"
instead of retrying.

Reported-By: DJ-RK