Warn if both HLSL and Assembly versions of a given shader exist
If both Assembly and HLSL versions of the same shader are found in
ShaderFixes 3DMigoto will have to choose one over the other, but
although we do have a policy on prioritising the HLSL shader over
assembly, this is not the correct thing to do.

There's a few ways we could have got into this situation:

- A shaderhacker might decide to switch a shader from HLSL to Assembly
  or vice versa. Prioritising the HLSL shader is wrong in this case -
  the correct course of action is for the shaderhacker to remove one of
  the shaders.

- An end user might be updating a fix by extracting the new one over an
  old one. If the old version included a HLSL version of a given shader
  and the new one includes an assembly version (or vice versa) it is
  again wrong to prioritise the HLSL shader - the correct course of
  action is for the user to use uninstall.bat to remove the old version
  of the fix then install the new version.

- A shaderhacker might dump a shader via hunting that was already in
  ShaderFixes as assembly. Again, prioritising the HLSL version is the
  wrong thing to do - the correct course of action is to touch the
  Assembly version that has already been dumped instead of dumping the
  HLSL version, or alternatively the shaderhacker could choose which
  version to use and remove the other.

- A shaderhacker might decide to switch a shader from HLSL to Assembly
  and remove the .txt file but neglect to remove the cached .bin file.
  While we now check the timestamp of the .bin file with the
  corresponding .txt file, we only compare that for shaders generated
  from the same type - a stale _replace.bin file compiled from HLSL
  could still trump a newer assembly text file. In this case we probably
  could make a policy decision, but it is equally valid to just have the
  shaderhacker remove the .bin file.

In each of these cases it is really not 3DMigoto's place to decide what
to do, so instead just detect when the scenario has arisen and issue a
dire warning on the overlay.

If this is detected on initial load we assume it might be an end user
and ask them to run uninstall.bat and reinstall the fix, but if it is
detected on an F10 reload or copy on mark we assume it is a shaderhacker
and just notify them of which files are in conflict.