Set load_library_redirect=2 by default.
Change default to =2 because of noting problems with multiple
games where having it set to zero is failing.  Too many examples
where it has been failing.

=0 is needed for proxy loading and typically when using other
injectors/overlays.  But can be done on an as needed basis,
because those are more rare.

This puts it back to 1.2.x standard behavior.