Add Injector project to inject 3DMigoto into any application
This is a pretty simple injector that can inject 3DMigoto into any
running application via SetWindowsHookEx.

The 3DMigoto DLL must currently be named "d3d11.dll" - this means that
it will "just work" in many games (such as Unity/Subnautica), however
this may not be reliable (doesn't seem to work in RoTTR under UWP), so
this may not be final - we may instead name it something else and have
it hook into the d3d11.dll exports once injected.

This is set up with the intention of being used to inject 3DMigoto into
Windows store apps (but can be used to inject into anything), so it has
UAC UIAccess = true. This means that the resulting binary must be
digitally signed, which is currently not done automatically by the build
system.

If UWP injection is not desired, just turn this option off to enable
injection without signing into any other application.

If UWP injection is desired, a self-signed signing certificate may be
generated with this Powershell command (alternatively we may be able to
use makecert on older versions of Windows):

    New-SelfSignedCertificate -CertStoreLocation cert:\currentuser\my `
    -Subject "CN=3DMigoto" `
    -FriendlyName "3DMigoto Code Signing Certificate" `
    -KeyUsage DigitalSignature `
    -Type CodeSigningCert

Certmgr.exe ("manage user certificates" is fine - no need to "manage
computer certificates") should then be used to move/copy this
certificate into "Trusted Root Certificate Authorities" (copying seems a
little easier, so the below SignTool command will just work in that
case - it should be possible to tell SignTool to use the certificate in
trusted CAs, but I can't work out how).

The executable then needs to be signed. From a Visual Studio Developer
terminal, run:

    SignTool sign /debug /v /s "my" /n "3DMigoto" /t http://timestamp.verisign.com/scripts/timestamp.dll Injector.exe

This may not be the final way we do this signing - if we were a
commercial entity we would want a signing certificate from a CA already
trusted by Windows, but as we are not that is a little too expensive, so
we may opt to use an installer to install this certificate on the user's
machines, and maybe we could make our own root certificate that has
signed each individual developer's code signing certificates rather than
sharing a private certificate between all developers.

Additionally, if this injector is to be used for Windows Store Apps, it
must be installed somewhere under c:\program files (x86), because of a
Windows security policy. This is not required if injecting into other
applications.