Implement CreateDXGIFactory2
This function can be used on Windows 8.1+, so we need to support it to
ensure we can get access to the swap chain of any game that uses it.

It is known to be used in the Windows Store version of Rise of the Tomb
Raider, but is not specific to the Windows Store, so we need to support
it regardless.

Note that this function is not in the Windows 8.0 SDK, but since we are
locating it via GetProcAddress and hooking it and it doesn't require any
new types we can still support it.

We do have to be careful not to refer to it statically though (which is
only possible on the vs2015 branch with a newer Windows SDK), as doing
so would introduce a link time dependency on a newer dxgi.dll and
prevent us being loaded on Windows 7.