DXGI: Implement CreateSwapChainForCoreWindow and Composition
The CoreWindow variant is used by Windows Store apps, so we need it to
get access to the swap chain for those. We might not care about the
DirectComposition variant, but the code is identical, so just go ahead
and implement it anyway.

Neither of these have a hwnd to capture, so we won't have mouse support
yet, though the CoreWindow version does have a hWindow that we will be
able to get a hWnd from.

The code is refactored to use as much code in common with
CreateSwapChainForHwnd, and further refactoring with the original
CreateSwapChain may be possible.

The override & wrap helper functions for the original CreateSwapChain
have been moved up in the file to keep them next to their near identical
Factory2 variants.