Fix overlay when GetModuleHandle("d3d11.dll") doesn't return 3DMigoto
GetModuleHandle("d3d11.dll") is not guaranteed to return a handle to
3DMigoto since there is at least one other d3d11.dll in the address
space (although in practice it has seemed pretty reliable), and when we
are injected into Windows Store apps we might not even be named
d3d11.dll at all, in which case it will definitely fail.

Switch to getting the module handle by address to guarantee we get a
handle to ourselves to fix the overlay in Windows Store apps.
Alternatively, we could save off the handle we get passed in DllMain.