Limit "overriding NVAPI wrapper failed" warnings to one
If we are shipped without our nvapi wrapper the log becomes very noisy
with this message. We probably do want this message in the log so we
can see why certain things may not be working, but we don't need it
flooding the log, so limit it to printing out at most one time in the
log.