Allocate Thread Local Storage
This gives us a structure we can use for any per-thread variables we may
want in the future. It consumes a single TLS slot for the entire
structure, giving us the freedom to use it for as much as we need
without worrying about exhausting the limited resource of TLS slots.

There is also a "thread_local" storage class specifier in C++11 that
would do the same thing (but be vastly simpler), however MSDN warns that
it can interfere with delay loading DLLs, so without knowing the details
err on the side of caution and use the Win32 TLS API.