Hooking: Add protection against inverted call chains due to hooking
If we hook a DirectX function we risk an inverted call chain where
DirectX (or a third party tool sitting between us and DirectX) can
unexpectedly call back into us, causing all sorts of problems.

One such problem that we have already seen, is that
D3D11CreateDeviceAndSwapChain is implemented by calling
IDXGIFactory::CreateSwapChain, so if we have hooked the later function,
calling the former will unexpectedly call back into us.

To resolve this, we re-implemented D3D11CreateDeviceAndSwapChain
ourselves, but in doing so introduced two new problems:

1. SpecialK implements D3D11CreateDevice by calling
   D3D11CreateDeviceAndSwapChain, and could call back into us.

2. DirectX implements D3D11CreateDevice by calling
   D3D11CreateDeviceAndSwapChain, and could call back into us if we also
   hook D3D11CreateDeviceAndSwapChain (which we don't do at the moment,
   but will want to in the future to allow our DLL to be called
   something other than d3d11.dll)

We worked around problem 1 by splitting our D3D11CreateDevice
implementation into a separate function, but this doesn't help for case
2, and depending on the order that we and other tools get loaded and
what hooking / wrapping techniques are in use by all tools there could
be other potential cases where this may fall short as well (for example,
SpecialK could have been sitting between us and DirectX, and call back
into us via any DirectX function it might call that we had hooked).

Really this is reflective of one of the fundamental problems with
hooking, but since we can't avoid hooking without accepting increased
complexity, sacrificing functionality and compatibility with certain
games, and since we can't expect every other tool on the planet to ditch
hooking as well, we will just have to suck it up and come up with
strategies to deal with the fallout.

One such strategy we could employ is to set a flag whenever we call into
external code, and check if that flag is set whenever we get called by
external code, and if so call straight through to the original function
with no further processing. This can solve the cases where code we call
can call back into us, but will not solve the cases where we call a
function we expect to be internal, but which has actually been hooked by
a third party - for that we will need strategies like the unhookable
HackerCreateDevice, or avoiding re-implementing any function in terms of
others that may have been hooked.

However, employing this flag strategy when calling all external code is
a bit extreme - we technically should do this if we want hooking to be
safe, but in 99% of cases it is unnecessary, and a little insane to add
it everywhere.

So, instead we will just protect known or suspected problem cases, and
expand the coverage as we find it necessary. This commit adds this
protection for calling and being called on the following functions:

- D3D11CreateDevice
- D3D11CreateDeviceAndSwapChain
- IDXGIFactory::CreateSwapChain
- IDXGIFactory2::CreateSwapChainForHwnd

This solves known issues where DirectX can call into us via our hooks on
D3D11CreateDeviceAndSwapChain and IDXGIFactory::CreateSwapChain.