HackerCreateDevice: Split out device & context wrapping into separate function
The act of wrapping the device and context is logically separate from
the rest of the processing we do in D3D11CreateDevice, so split it out
into a separate function.

Additionally, this means if we were to restore the original
implementation of D3D11CreateDeviceAndSwapChain (thereby avoiding the
need for the unhookable HackerCreateDevice entirely) and do a similar
refactoring job with CreateSwapChain, it would only be a fairly short
function instead of the few pages of code it used to be - it was only
ever complicated because the code had not been refactored properly.