CreateSwapChain: Split up into several functions
The act of collecting the resolution from the swap chain description and
overriding it is logically separate from calling through to the
original, and separate from wrapping the swap chain, so refactor these
out into helper functions.

Annoyingly there are a few too many small differences with the factory2
versions to reuse these as common code there (we could break them up
even smaller), but it's still cleaner to split these out, and if we ever
restore D3D11CreateDeviceAndSwapChain to call through to the original,
these will allow it to be much smaller than the old implementation.