Frame Analysis: Correct draw call numbers in deferred contexts
The deferred contexts frame analysis log file contained incorrect draw
call numbers, as these were shared with the immediate context and never
incremented in the deferred contexts.

The analyse_frame global had two meanings - it indicated whether frame
analysis was currently being performed, and held the current draw call
number. This splits the two meanings out, so the draw call number is now
per-context, leaving the global analyse_frame only indicating frame
analysis is currently active.

This global meaning of analyse_frame is still potentially inaccurate,
since there is no constraint that says a deferred context must be
written to only during the same frame as the immediate context where it
is eventually used, but it is difficult to retrospectively set it in
deferred contexts if it is not, and the other alternative is to delay
frame analysis by one frame to propagate it to all the deferred
contexts, but even then it may still not be accurate in all cases.
So far this has not caused any significant issues, and analyse_options =
hold can capture multiple frames to make up for this.

The draw call number is now incremented on draw calls from deferred
contexts as well, though these still do not dump resources.