Full Screen: Add new mechanism to instantly force full screen
Forcing full screen in games that don't have a native option for it is a
bit of a black art. This adds a new key binding which can be used to try
to instantly force full screen when pressed.

This is necessary for DOAXVV as the launcher must run in a window or
none of the buttons work, and full_screen=1 interferes with this, and
even if that is worked and the main game has launched and in full
screen, it will not return to full screen after an alt+tab. This gives
us a way to force full screen only when we need to.