Frame Analysis: Add experimental option to dump from deferred contexts
Frame Analysis originally only dumped render and depth targets after
every draw call, which doesn't really work for deferred contexts since
the resources aren't updated until later, and deferred contexts cannot
be used to stage resources to the CPU, so because of those reasons it
just bailed for any deferred contexts.

Since then it has gained support for dumping other types of resources,
and these may still be useful to dump provided that they are not updated
in the deferred context, such as static textures, constant buffers,
vertex and index buffers. This can be especially useful if we want to do
texture filtering in a deferred context, since we generally are
filtering on static textures that (usually) are not updated on the GPU.

This adds experimental support to create frame analysis dumps from
deferred contexts. It dumps these resources into a subdirectory of the
main frame analysis directory so they don't clash with resources dumped
from the immediate context.

This currently uses the immediate context to dump the resources, which
seems to work fine in DOAXVV, but could possibly cause issues in other
games that perform more work in both if there are any threading issues
with issuing commands to copy resources, map or unmap on the immediate
context from another thread.

There may also be some threading issues with the NVAPI call used to set
the surface creation mode to dump stereo resources, since the call
appears to be global, not per-context.

It's not clear if either of these will prove to be a problem in
practice, so to be safe add it as an option that must be enabled by the
user to eliminate the risk in the normal case, and so that they
acknowledge the limitations and potential problems when they enable it.