Frame Analysis: Refactor resource staging out of DumpStereoResources
For the work to make frame analysis dumping some orders of magnitude
faster we will need to stage all Texture2D resources manually (i.e. not
leaving it to DirectXTK), even if not dumping them as stereo, so
refactor this out of DumpStereoResources to enable us to use it on mono
resources.