Frame Analysis: Use symbolic links to de-duplicate dumped files
Frame analysis dumping can be notoriously slow and consume enormous
amounts of disk space. We do have strategies to cope with this, such as
setting up triggers to dump only part of a frame, but the fact is that
many of the dumped files are identical and we could do a lot better if
we only dumped out each once.

This re-hashes the resources just before they are dumped, and uses that
to create a filename for each unique dumped file, so that attempting to
dump the same resource multiple times will only write the file out once
if it has not been modified in the meantime. This uses the same hashing
scheme as our texture filtering, and we specifically use the original
resource description to calculate it (as opposed to the resource
description of the staging resource) to increase the odds that the hash
will match that used in 3DMigoto - but this is not guaranteed, as the
resource may have been altered on the GPU (e.g. rendered to), and in
some cases the resource description that DirectX returns to us may not
be identical to the one used to create the resource.

These unique files are then symbolic linked into the main frame analysis
dump directory so that frame analysis will continue to operate
functionally as before.