Frame Analysis: Add dump_info option to dump resource descriptions
We often want to check the resource description matching a given hash
(e.g. to note down what the hash means, or to use a fuzzy texture
override), which we currently have to get from ShaderUsage.txt.

Now that we have the dedupe folder though, the files contained within
can be a little more disconnected since the hash in that directory is
not guaranteed to match the texture filtering hash, but we still might
want to get the resource info for these, so the easiest way is to dump
it to a file next to the dumped resources.

Move the resource description string formatting routines into
ResourceHash.cpp and use them to generate a .nfo file next to the
dumped resources.