Frame Analysis: Zero out padding instead of skipping
This helps many of the dedupe hashes to match the texture hashes - about
half of the HUD textures that should match in DOAXVV are now matching.

Further investigation will be required to determine why the remaining
hashes are still not matching - possible theories include SysMemPitch
and RowPitch not matching, or garbage in the padding bytes at the time
the resource was created.