Frame Analysis: Fix inconsistent dedupe hashes
Various static textures were being dumped out with inconsistent hashes
in the dedupe folder, meaning it wasn't quite as de-duplicated as it was
supposed to be, and that the hashes would often not match the texture
hashes even when the conditions were right for them to do so.

This is because the padding between rows in a texture can contain
garbage from vram, and we need to skip over them to ensure we are only
hashing the actual texture.

This skips over the padding, which makes the hashes consistent, but does
not help match the texture hashes (except in cases where there is not
padding). For that we'd probably need to replace the padding with zeroes
instead, but there may be some other complexity there, such as
SysMemPitch used when creating the resource possibly not matching the
RowPitch used by DirectX or garbage in the padding bytes when creating
the resource - more investigation will be required.

Ideally, we should always have been skipping the padding bytes, but
DirectX doesn't exactly make this easy.