Frame Analysis Deduplication: Add windows shortcut and move file fallbacks
If a game is running off a FAT32 partition than both symlinking and hard
linking will fail since neither are supported by the filesystem, so in
that case we fall back to creating a Windows Shortcut file. These are
pretty garbage, but they don't require filesystem support.

In the event that also fails for some reason, as a final fallback just
give up and move the deduplicated file in place (and further dumps would
do the same).