Frame Analysis: Move ShaderOverride triggers into command list
This moves the analyse_options handling from ShaderOverrides and
TextureOverrides into the command list. This allows more flexibility in
how frame analysis options can be changed during a frame, and fulfils
the long standing TODO items to allow frame analysis to be triggered
from any texture - but with the caveat that checktextureoverride is
necessary for those.

This further decouples FrameAnalysisContext from HackerContext, as it
removes the final dependency on mCurrentRenderTargets &
mCurrentDepthTarget from the frame analysis code.