Command List / Frame Analysis: Add "dump" command
This command will have frame analysis dump the specified resource if it
is currently active, possibly specifying custom options. It is useful if
only a specific resource needs to be dumped without dumping all other
resources of the same class (e.g. dump = mono ps-t0), or to dump a
resource at a specific point in time (e.g. before and after the draw
call) or to dump a custom resource that frame analysis normally cannot
see (e.g. dump = ResourceDepthBuffer).