Frame Analysis: Give buffer .txt filenames extra info for full uniqueness
The buffers dumped as text files include metadata not found in the
binary files, and may not include part of the buffer excluded by an
offset. Since these can be many and varied we need to include them in
the filename to make sure we don't falsely deduplicate buffers that are
binary identical, but vary by these.

This adds the metadata into the deduped text filename so that multiple
text files can exist for a single buffer if their metadata is different.

Buffer binary filenames are unchanged, since the metadata does not come
into consideration for those.