Frame Analysis: Replace DUMP_ON_XXXXXX with bump_buf_bin/txt
DUMP_ON_XXXXXX was essentially a hack to get dump on update/unmap to
work regardless of what kind of resource it was, however this strategy
could not use the vertex/index buffer specific dump routines as they
were not in this mask, and would use the general routine instead. This
could be a problem for the command list dump command, as a user might
specify something more informative.

Remove this double meaning, and use DUMP_XXX_XX and DUMP_XXX_XX_MASK to
select default dump options and check which dump options are set.

Add bits specific to dumping buffers as binary/text that we can select
independently, and remove DUMP_ON_XXXXXX and the matching bits in all
masks.

This is a stepping stone commit to clean up the existing bits before
completely separating out the concepts of what binds we are dumping and
what formats we are dumping.