Frame Analysis: Separate bind and format selection flags & cleanup
This splits options like "dump_rt_jps" into separate "dump_rt" and
"jps" options. This loses the ability to do things like "analyse_options
= dump_rt_jps dump_tex_dds", since this will now be a combination of all
options ("dump_rt dump_tex jps dds"), however now that we can dump
specific resources from the command list this is not as big of a deal,
since we can specify individual options there.

Overall, while a fairly significant change with high potential for
regressions,  this is a simplification to the flags and code, and should
make this easier to maintain into the future. It also buys us a few bits
for additional options before we have to increase the size of
FrameAnalyisOptions (and likely having to sacrifice the enum class in
doing so).