Frame Analysis: Delay dumping buffers from deferred contexts
For Frame Analysis to be accurate when used in a deferred context we
actually want to delay reading the resources being dumped on the CPU
until we are back in the immediate context.

This adds a new experimental option to do this - it will go through the
motions to set the commands to perform the reverse stereo blit and stage
the texture to the CPU, but places it in a data structure instead of
reading or dumping the resource. When FinishCommandList is called, the
data structure is detached from the deferred context and placed in a
global lookup map. Finally, when ExecuteCommandList is called back on
the immediate context, the data structure is fetched from the global
lookup map and the deduplication and dumping stages are performed.

This should only be used with relatively small dumps, since they need to
remain in memory for some time, and the de-duplication step is performed
too late to minimise that.

Initial testing suggests that the reverse stereo blit does not work with
these.